Fix up image storage so everything is done by scanlines
The image is now has shape image[Y][X].  This means that the natural
loop over pixels is now scanline-wise, which is easier to think
about.  In addition separate the number of pixels in the X and Y
dimensions so that you can look at non-square domains with the correct
aspect ratio.